# Conflicts:
#	src/main/kotlin/inc/evil/d17map/controllers/EquipmentController.kt
#	src/main/kotlin/inc/evil/d17map/controllers/ReservationController.kt
#	src/main/kotlin/inc/evil/d17map/services/EquipmentService.kt
#	src/main/kotlin/inc/evil/d17map/services/ReservationService.kt

+++ src/main/kotlin/inc/evil/d17map/services/EquipmentService.kt
@@ -2,6 +2,7 @@ package inc.evil.d17map.services
 
 import inc.evil.d17map.dtos.EquipmentRequest
 import inc.evil.d17map.dtos.EquipmentResponse
+import inc.evil.d17map.EquipmentNotFoundException
 import inc.evil.d17map.entities.Equipment
 import inc.evil.d17map.findOne
 import inc.evil.d17map.repositories.EquipmentRepository
@@ -22,6 +23,9 @@ class EquipmentService(val equipmentRepository: EquipmentRepository) {
     }
 
     fun removeEquipment(equipmentId: UUID) {
+        if (!equipmentRepository.existsById(equipmentId)) {
+            throw EquipmentNotFoundException(equipmentId)
+        }
         val equipment = equipmentRepository.findOne(equipmentId)!!
         equipmentRepository.delete(equipment)
     }
